Fix pos tests.
Disabled separate compilation as now it does not work anymore.
The issue is that the class for the phantom object is neber emmited
as it is eliminated. The solution will be to emmit the phantom class
as an empty class/object with private constructors. THis way the tasty
will be in that file.